+++ src/components/KymaCompanion/api/getChatResponse.ts
@@ -5,18 +5,17 @@ import {
   MessageChunk,
 } from '../components/Chat/Message/Message';
 import { HttpError } from './error';
-
-const ERROR_RETRY_TIMEOUT = 1_000;
-const MAX_ATTEMPTS = 3;
+import {
+  handleChatErrorResponseFn,
+  handleChatResponseFn,
+  retryFetch,
+} from 'components/KymaCompanion/api/retry';
 
 interface ClusterAuth {
   token?: string;
   clientCertificateData?: string;
   clientKeyData?: string;
 }
-
-type handleChatResponseFn = (chunk: MessageChunk) => void;
-type handleChatErrorResponseFn = (errResponse: ErrResponse) => void;
 
 type GetChatResponseArgs = {
   query: string;
@@ -45,7 +44,7 @@ const fillAuthHeaders = (
     throw new Error('Missing authentication credentials');
   }
 };
-const prepareBasicHeaders = (
+const createBasicHeaders = (
   certificateAuthorityData: string,
   clusterUrl: string,
   sessionID: string,
@@ -59,94 +58,6 @@ const prepareBasicHeaders = (
   };
   return headers;
 };
-export default async function getChatResponse({
-  query,
-  namespace = '',
-  resourceType,
-  groupVersion = '',
-  resourceName = '',
-  sessionID,
-  handleChatResponse,
-  handleError,
-  clusterUrl,
-  clusterAuth,
-  certificateAuthorityData,
-}: GetChatResponseArgs): Promise<void> {
-  const { backendAddress } = getClusterConfig();
-  const url = `${backendAddress}/ai-chat/messages`;
-  const payload = {
-    query,
-    namespace,
-    resourceType,
-    groupVersion,
-    resourceName,
-  };
-  const headers = prepareBasicHeaders(
-    certificateAuthorityData,
-    clusterUrl,
-    sessionID,
-  );
-  fillAuthHeaders(headers, clusterAuth);
-
-  const fetchWrapper = async function(
-    handleResponse: handleChatResponseFn,
-    handleError: handleChatErrorResponseFn,
-  ): Promise<boolean> {
-    return fetchResponse(
-      url,
-      headers,
-      JSON.stringify(payload),
-      sessionID,
-      handleResponse,
-      handleError,
-    );
-  };
-
-  await retryFetch(fetchWrapper, handleChatResponse, handleError);
-}
-
-async function retryFetch(
-  fetchFn: {
-    (
-      handleResponse: handleChatResponseFn,
-      handleError: handleChatErrorResponseFn,
-    ): Promise<boolean>;
-  },
-  handleChatResponse: handleChatResponseFn,
-  handleError: handleChatErrorResponseFn,
-) {
-  let finished = false;
-  for (let i = 0; i < MAX_ATTEMPTS; i++) {
-    const handleErrWrapper = (errResponse: ErrResponse) => {
-      errResponse.maxAttempts = MAX_ATTEMPTS;
-      errResponse.attempt = i + 1;
-      handleError(errResponse);
-      console.debug('2: Finished handling the error', errResponse);
-    };
-    const handleChatResponseWrapper = (chunk: MessageChunk) => {
-      handleChatResponse(chunk);
-      console.debug('2: Finished reading the answer');
-    };
-
-    finished = await fetchFn(handleChatResponseWrapper, handleErrWrapper);
-
-    console.debug('3: Fetch Done');
-    if (!finished) {
-      await new Promise(resolve => setTimeout(resolve, ERROR_RETRY_TIMEOUT));
-    } else {
-      console.debug('DONE');
-      finished = true;
-      break;
-    }
-  }
-  if (!finished) {
-    handleError({
-      message:
-        "Couldn't fetch response from Kyma Companion because of network errors.",
-      type: ErrorType.FATAL,
-    });
-  }
-}
 
 async function fetchResponse(
   url: RequestInfo | URL,
@@ -227,4 +138,50 @@ async function readChunk(
       });
       console.error('Error reading stream:', error);
     });
+}
+
+export default async function getChatResponse({
+  query,
+  namespace = '',
+  resourceType,
+  groupVersion = '',
+  resourceName = '',
+  sessionID,
+  handleChatResponse,
+  handleError,
+  clusterUrl,
+  clusterAuth,
+  certificateAuthorityData,
+}: GetChatResponseArgs): Promise<void> {
+  const { backendAddress } = getClusterConfig();
+  const url = `${backendAddress}/ai-chat/messages`;
+  const payload = {
+    query,
+    namespace,
+    resourceType,
+    groupVersion,
+    resourceName,
+  };
+  const headers = createBasicHeaders(
+    certificateAuthorityData,
+    clusterUrl,
+    sessionID,
+  );
+  fillAuthHeaders(headers, clusterAuth);
+
+  const fetchWrapper = async function(
+    handleResponse: handleChatResponseFn,
+    handleError: handleChatErrorResponseFn,
+  ): Promise<boolean> {
+    return fetchResponse(
+      url,
+      headers,
+      JSON.stringify(payload),
+      sessionID,
+      handleResponse,
+      handleError,
+    );
+  };
+
+  await retryFetch(fetchWrapper, handleChatResponse, handleError);
 }
